+++ src/main/java/com/taxamo/client/model/InputTransactionUpdate.java
@@ -76,8 +76,7 @@ public class InputTransactionUpdate {
   private String supplyDate = null;
   /* Transaction lines. */
   private List<InputTransactionLine> transactionLines = new ArrayList<InputTransactionLine>();
-  /* Order date in yyyy-MM-dd format, in merchant's timezone. If provided by the API caller, no timezone conversion is performed.
-   Default value is current date and time. When using public token, the default value is used. */
+  /* Order date in yyyy-MM-dd format, in merchant's timezone. If provided by the API caller, no timezone conversion is performed. Default value is current date and time. When using public token, the default value is used. */
   private String orderDate = null;
   public String getInvoiceDate() {
     return invoiceDate;